Fixed problem where if a finger paused the gesture recognition but didn't result in a sound being played, the other fingers would not play. Now if the first finger that triggers a gesture doesn't play a sound, the other fingers are checked for position.
Implemented MIDI like messages to pass to the audio classes. Messages include velocity (0-127), which drum is hit left/right, and how far away from the drum center it is (0-127) with 0 being the center and 127 the edge.
Also changed to using JSyn rather than Beads as found that results were faster after experiments.

+++ MidiMessage.java
@@ -13,8 +13,8 @@ public class MidiMessage
   //
   /**************************************************************************************************/
   
-  private int velocity;   // Velocity of the message
-  private int note;       // The note to play 
+  private byte velocity;   // Velocity of the message
+  private byte note;       // The note to play 
   private TablaDrum drum; // Which drum: left/right/neither
   
   
@@ -42,7 +42,7 @@ public class MidiMessage
     }
     else
     {
-      this.velocity = velocity;
+      this.velocity = (byte) velocity;
     }
     
     
@@ -56,7 +56,7 @@ public class MidiMessage
     }
     else
     {
-      this.note = note;
+      this.note = (byte) note;
     }
     
     this.drum = drum;
@@ -76,7 +76,7 @@ public class MidiMessage
    * @return int the veloicity for the note between 0-127
    */
    
-  public int getVelocity()
+  public byte getVelocity()
   {
     return velocity;
   }
@@ -87,7 +87,7 @@ public class MidiMessage
    * @return int the note between 0-127
    */
    
-  public int getNote()
+  public byte getNote()
   {
     return note;
   }
